+++ test/modify-schema.test.ts
@@ -66,6 +66,32 @@ describe('modifySchema', () => {
           },
         },
       },
+      pet_clinics: {
+        'title': 'Pet clinics',
+        'type': 'array',
+        'x-jsf-presentation': {
+          inputType: 'group-array',
+        },
+        'items': {
+          type: 'object',
+          properties: {
+            name: {
+              'title': 'name',
+              'type': 'string',
+              'x-jsf-presentation': {
+                inputType: 'select',
+              },
+            },
+            address: {
+              'title': 'address',
+              'type': 'string',
+              'x-jsf-presentation': {
+                inputType: 'text',
+              },
+            },
+          },
+        },
+      },
     },
     'required': ['has_pet'],
     'x-jsf-order': ['has_pet', 'pet_name', 'pet_age', 'pet_fat', 'pet_address'],
@@ -336,6 +362,166 @@ describe('modifySchema', () => {
       ])
     })
 
+    it('replaces values in existing array item properties under config.fields.<arrayPath>.items.properties', () => {
+      const schemaPetWithArrayItemFields = {
+        ...schemaPet,
+        properties: {
+          ...schemaPet.properties,
+          meta: {
+            properties: {
+              description: 'Metadata of the pet',
+              tagNumbers: {
+                title: 'Tag History',
+                type: 'array',
+                items: {
+                  title: 'Tag Details',
+                  type: 'object',
+                  required: ['date', 'code'],
+                  properties: {
+                    date: {
+                      type: 'string',
+                      title: 'Date',
+                    },
+                    code: {
+                      type: 'string',
+                      title: 'Code',
+                    },
+                  },
+                },
+              },
+            },
+          },
+        },
+      }
+
+      const result = modifySchema(schemaPetWithArrayItemFields, {
+        fields: {
+          'meta.tagNumbers': {
+            items: {
+              title: 'Tag Entry',
+              properties: {
+                date: {
+                  title: 'Date applied',
+                  type: 'string',
+                },
+                code: {
+                  title: 'Tag code',
+                  minLength: 1,
+                },
+              },
+            },
+          },
+        },
+      })
+
+      expect(result.schema.properties?.meta).toEqual({
+        properties: {
+          description: 'Metadata of the pet',
+          tagNumbers: {
+            title: 'Tag History',
+            type: 'array',
+            items: {
+              title: 'Tag Entry',
+              type: 'object',
+              required: ['date', 'code'],
+              properties: {
+                date: {
+                  type: 'string',
+                  title: 'Date applied',
+                },
+                code: {
+                  type: 'string',
+                  title: 'Tag code',
+                  minLength: 1,
+                },
+              },
+            },
+          },
+        },
+      })
+      expect(result.warnings).toEqual([])
+    })
+
+    it('replace array item fields under config.fields.<arrayPath>.items.properties that dont exist gets ignored', () => {
+      const schemaPetWithArrayItemFields = {
+        ...schemaPet,
+        properties: {
+          ...schemaPet.properties,
+          meta: {
+            properties: {
+              description: 'Metadata of the pet',
+              tagNumbers: {
+                title: 'Tag History',
+                type: 'array',
+                items: {
+                  title: 'Tag Details',
+                  type: 'object',
+                  required: ['date', 'code'],
+                  properties: {
+                    date: {
+                      type: 'string',
+                    },
+                    code: {
+                      type: 'string',
+                    },
+                  },
+                },
+              },
+            },
+          },
+        },
+      }
+
+      const result = modifySchema(schemaPetWithArrayItemFields, {
+        fields: {
+          'meta.tagNumbers': {
+            items: {
+              properties: {
+                // fields that don't exist in the original schema are ignored
+                vetNumber: {
+                  type: 'number',
+                },
+              },
+            },
+          },
+        },
+      })
+
+      expect(result.schema.properties?.meta).toEqual({
+        properties: {
+          description: 'Metadata of the pet',
+          tagNumbers: {
+            title: 'Tag History',
+            type: 'array',
+            items: {
+              title: 'Tag Details',
+              type: 'object',
+              required: [
+                'date',
+                'code',
+              ],
+              properties: {
+                date: {
+                  type: 'string',
+                },
+                code: {
+                  type: 'string',
+                },
+                // vetNumber is not added
+              },
+            },
+          },
+        },
+      })
+
+      expect(result.warnings).toEqual([
+        {
+          type: 'FIELD_TO_CHANGE_NOT_FOUND',
+          message: 'Changing field "vetNumber" was ignored because it does not exist.',
+        },
+      ])
+    })
+
     it('replace all fields', () => {
       const result = modifySchema(schemaPet, {
         allFields: (fieldName, fieldAttrs) => {
@@ -473,6 +659,63 @@ describe('modifySchema', () => {
         },
       })
     })
+
+    it('replace field attrs inside array items', () => {
+      const result = modifySchema(schemaPet, {
+        fields: {
+          'pet_clinics.items.name': {
+            title: 'Clinic name',
+          },
+        },
+      })
+
+      expect(result.schema).toMatchObject({
+        properties: {
+          pet_clinics: {
+            items: {
+              properties: {
+                name: {
+                  title: 'Clinic name',
+                },
+              },
+            },
+          },
+        },
+      })
+    })
+
+    it('replace field attrs inside array items using allFields', () => {
+      const result = modifySchema(schemaPet, {
+        allFields: (fieldName, fieldAttrs) => {
+          if (fieldAttrs['x-jsf-presentation']?.inputType === 'select') {
+            return {
+              'title': 'Clinic name',
+              'x-jsf-presentation': {
+                inputType: 'text',
+              },
+            }
+          }
+          return {}
+        },
+      })
+
+      expect(result.schema).toMatchObject({
+        properties: {
+          pet_clinics: {
+            items: {
+              properties: {
+                name: {
+                  'title': 'Clinic name',
+                  'x-jsf-presentation': {
+                    inputType: 'text',
+                  },
+                },
+              },
+            },
+          },
+        },
+      })
+    })
   })
 
   describe('supporting custom attributes', () => {
@@ -540,7 +783,7 @@ describe('modifySchema', () => {
         },
       })
 
-      // Assert all the other propreties are kept
+      // Assert all the other properties are kept
       expect(result.schema).toMatchObject(invoiceSchema)
 
       expect(result.schema).toMatchObject({